✨ feat(mobile/auth): rename log-in to login, create signup page, and update navigation paths
