--- apps/mobile/src/app/(auth)/log-in.tsx
@@ -1,3 +1,0 @@
-import { LoginPage } from "../../layers/pages/log-in";
-
-export default LoginPage;
