+++ apps/mobile/src/app/(auth)/login.tsx
@@ -1,0 +1,3 @@
+import { LoginPage } from "../../layers/pages/login";
+
+export default LoginPage;
