--- apps/mobile/src/layers/pages/log-in/ui/login.header.tsx
@@ -1,3 +1,0 @@
-import { WaveHeader } from "../../../widgets/layout";
-
-export const LoginPageHeader = () => <WaveHeader title="Log In" enableBackButton />;
