+++ apps/mobile/src/layers/pages/login/ui/login.header.tsx
@@ -1,0 +1,3 @@
+import { WaveHeader } from "../../../widgets/layout";
+
+export const LoginPageHeader = () => <WaveHeader title="ログイン" enableBackButton />;
